+++ CHANGELOG.md
@@ -17,6 +17,7 @@ The client here will eventually be released as "spython" (and eventually to
 singularity on pypi), and the versions here will coincide with these releases.
 
 ## [master](https://github.com/singularityhub/singularity-cli/tree/master)
+ - fixing bug that inspect does not honor quiet (0.0.65)
  - refactor recipe parsers, writers, and base (0.0.64)
    - paths for files, add, copy, will not be expanded as it adds hardcoded paths
  - oci state fixes and added Client.version_info() (0.0.63)